Updated vendored deps
Using github.com/docker/libkv, which is unmaintained, instead of the one
maintained by the original author github.com/abronan/libkv as it has
references to coreos branch.

Fixed directly in vendor a bug in kv/etcd that introduces and error in keys.generated.go

We just delete the file that is rebuilt when build or run getconf:

    $ rm ../vendor/github.com/coreos/etcd/client/keys.generated.go

It is fixed in coreos/etcd master [issues/8715](https://github.com/coreos/etcd/issues/8715)
so it will be available in next etcd client release.